# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Secure Gift Card Manager API
  version: 0.1.0
  description: >
    Initial API contract for the local MVP. The implementation must keep this
    file in sync with backend routes and QA requirement IDs.
servers:
  - url: http://localhost:3001
    description: Local API server
...
paths:
...
  /api/deals:
    get:
      tags: [Deals]
      summary: List deals
      operationId: listDeals
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Deals page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Deals]
      summary: Create a deal, optionally with cards
      operationId: createDeal
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealRequest"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/deals/{dealId}:
    parameters:
      - $ref: "#/components/parameters/DealId"
    get:
      tags: [Deals]
      summary: Deal detail with associated cards
      operationId: getDeal
      responses:
        "200":
          description: Deal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Deals]
      summary: Update deal metadata with optimistic concurrency
      operationId: updateDeal
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealRequest"
      responses:
        "200":
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/deals/{dealId}/archive:
    post:
      tags: [Deals]
      summary: Archive a deal without removing cards
      operationId: archiveDeal
      parameters:
        - $ref: "#/components/parameters/DealId"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Deal archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/deals/{dealId}/unarchive:
    post:
      tags: [Deals]
      summary: Restore an archived deal to the default list
      operationId: unarchiveDeal
      parameters:
        - $ref: "#/components/parameters/DealId"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Deal unarchived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    DealId:
      name: dealId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated or locked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated request failed CSRF, origin, or authorization checks
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: State conflict or duplicate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CardStatus:
      type: string
      enum: [available, reserved, in_use, sold, used_up, void]
    CardType:
      type: string
      enum: [merchant, prepaid]
    Network:
      type: string
      enum: [visa, mastercard, amex, discover, other]
    Card:
      type: object
      required:
        - id
        - accountId
        - brand
        - cardType
        - faceValueCents
        - remainingBalanceCents
        - purchaseCostCents
        - status
        - rowVersion
      properties:
        id:
          type: integer
        accountId:
          type: integer
        dealId:
          type: integer
          nullable: true
        brand:
          type: string
        cardType:
          $ref: "#/components/schemas/CardType"
        network:
          $ref: "#/components/schemas/Network"
        faceValueCents:
          type: integer
          minimum: 1
        remainingBalanceCents:
          type: integer
          minimum: 0
        purchaseCostCents:
          type: integer
          minimum: 0
        cardNumberLast4:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        expirationDate:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/CardStatus"
        reservedFor:
          type: string
          nullable: true
        reservedUntil:
          type: string
          nullable: true
        reservedNotes:
          type: string
          nullable: true
        latestSalePriceCents:
          type: integer
          nullable: true
        rowVersion:
          type: integer
    CardInput:
      type: object
      required: [brand, cardType, faceValueCents]
      properties:
        brand:
          type: string
        cardType:
          $ref: "#/components/schemas/CardType"
        network:
          $ref: "#/components/schemas/Network"
        faceValueCents:
          type: integer
          minimum: 1
        purchaseCostCents:
          type: integer
          minimum: 0
        cardNumber:
          type: string
        pin:
          type: string
        billingZip:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
...
    Deal:
      type: object
      required: [id, accountId, name, rowVersion]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        name:
          type: string
        source:
          type: string
          nullable: true
        purchaseDate:
          type: string
          nullable: true
        inputTotalCostCents:
          type: integer
          nullable: true
          minimum: 0
        notes:
          type: string
          nullable: true
        archivedAt:
          type: string
          nullable: true
        rowVersion:
          type: integer
    CreateDealRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        source:
          type: string
        purchaseDate:
          type: string
        totalCostCents:
          type: integer
          minimum: 0
        notes:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardInput"
    UpdateDealRequest:
      type: object
      properties:
        rowVersion:
          type: integer
          minimum: 1
        name:
          type: string
        source:
          type: string
          nullable: true
        purchaseDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    DealDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [deal, cards]
          properties:
            deal:
              $ref: "#/components/schemas/Deal"
            cards:
              type: array
              items:
                $ref: "#/components/schemas/Card"
    DealListResponse:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Deal"
...
    ListEnvelope:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
        page:
          type: object
          required: [limit, offset, total, hasMore]
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
            hasMore:
              type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, code, message]
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string